agregaado filtros anteriores para hacer pruebas de busqueda

+++ lib/models/movimientos_model.dart
@@ -1,7 +1,7 @@
 class Movimiento {
   late int id;
   late String name;
-  late double valor;
+  late int valor;
   late String fecha;
   late bool isTrue;
